+++ packages/pluggableWidgets/gallery-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- We added configurable selection count visibility and clear selection button label template for improved row selection management.
+
 ### Fixed
 
 - We fixed an issue where setting the gallery gap to 0 caused an offset, which made the bottom border of items to dissapear.